Fixed a suds.xsd.sxbasic.SchemaObject resolve() invalid cached value bug.
When called with parameter nobuiltin=True on a node whose type is a builtin, the
original implementation cached the builtin type as the result so a repeated
resolve() call returned an incorrect value.

Added related test case. Refactored the resolve() implementation into an
external function dealing with result caching and in internal worker function
doing the actual type resolution. Updated todo list.