update failing choice parameter tests
Choice parameter values must all be specified when using positional arguments
instead of expecting only one and using it as the value of the first element in
that choice group. At most one of those values may still be set to something
other than None.

Added more related test cases.

--HG--
branch : assure-that-all-parameters-are-consumed-1385966771465