add 'extra param' test for a choice containing an empty sequence
Part of the test has been marked as 'expected to fail' since suds library
currently does not fully support this scenario. More to the point, in such
cases the choice structure may not be correctly recognized as optional and
related 'extra param' error messages may state an incorrect required argument
count.

--HG--
branch : assure-that-all-parameters-are-consumed-1385966771465