upgrade shared test WSDL template to support using custom named operations
Prior to this change, the single operation defined in the template was always
named 'f'. Now this is used as the default operation name but the caller can
specify a custom name for it.

The operation's soap-action attribute changed so there would be no confusion
between it and the operation name.

--HG--
branch : assure-that-all-parameters-are-consumed-1385966771465