add unwrapped parameter ArgParser tests
The added test is run for only some of the possibly interesting scenarios. At
the moment, it tests only operations taking at most one input parameter, but
has been prepared with as a general enough test template so it can be run in
more complex scenarios as well.

--HG--
branch : assure-that-all-parameters-are-consumed-1385966771465