add an indirect_parametrize pytest plugin to the test suite
The new pytest plugin allows a clean & well integrated way of defining test
parametrizations with data constructed using a function instead of having to be
explicitly passed to a pytest.mark.parametrize() call.

An alternative would be to use a function constructing pytest.mark.parametrize()
arguments but the current solution causes less clutter when used and seems more
intuitive to someone used to using pytest.mark.parametrize().

--HG--
branch : assure-that-all-parameters-are-consumed-1385966771465